Finish A Star and Updated UI to switch search methods.
A star is now working as expected, though it wont be the best way to route the fastest way through the network.  The Heristic would need to be updated to account for the time spend on the paths better.  Looks like we'll be sticking with Diijsktra.  I updated the UI to switch between the two to show the difference in accuracy and speed.